WIP update re_ui_ex with LI2 and made `list_item_scope` a function
