+++ crates/re_ui/src/list_item2/mod.rs
@@ -1,12 +1,12 @@
 mod basic;
-mod container;
 mod debug;
 mod list_item;
+mod scope;
 
 pub use basic::*;
-pub use container::*;
 pub use debug::*;
 pub use list_item::*;
+pub use scope::*;
 
 struct ContentResponse<ContRet> {
     data: ContRet,
